MAINT: move module docstrings into __init__.py and remove info.py files.
