+++ scipy/cluster/__init__.py
@@ -1,8 +1,25 @@
-#
-# spatial - Distances
-#
+"""
+=========================================
+Clustering package (:mod:`scipy.cluster`)
+=========================================
 
-from info import __doc__
+.. currentmodule:: scipy.cluster
+
+:mod:`scipy.cluster.vq`
+
+Clustering algorithms are useful in information theory, target detection,
+communications, compression, and other areas.  The `vq` module only
+supports vector quantization and the k-means algorithms.
+
+:mod:`scipy.cluster.hierarchy`
+
+The `hierarchy` module provides functions for hierarchical and
+agglomerative clustering.  Its features include generating hierarchical
+clusters from distance matrices, computing distance matrices from
+observation vectors, calculating statistics on clusters, cutting linkages
+to generate flat clusters, and visualizing clusters with dendrograms.
+
+"""
 
 __all__ = ['vq', 'hierarchy']
