+++ scipy/fftpack/__init__.py
@@ -1,9 +1,94 @@
-#
-# fftpack - Discrete Fourier Transform algorithms.
-#
-# Created: Pearu Peterson, August,September 2002
+"""
+==================================================
+Discrete Fourier transforms (:mod:`scipy.fftpack`)
+==================================================
 
-from info import __all__,__doc__
+Fast Fourier Transforms (FFTs)
+==============================
+
+.. autosummary::
+   :toctree: generated/
+
+   fft - Fast (discrete) Fourier Transform (FFT)
+   ifft - Inverse FFT
+   fft2 - Two dimensional FFT
+   ifft2 - Two dimensional inverse FFT
+   fftn - n-dimensional FFT
+   ifftn - n-dimensional inverse FFT
+   rfft - FFT of strictly real-valued sequence
+   irfft - Inverse of rfft
+   dct - Discrete cosine transform
+   idct - Inverse discrete cosine transform
+
+Differential and pseudo-differential operators
+==============================================
+
+.. autosummary::
+   :toctree: generated/
+
+   diff - Differentiation and integration of periodic sequences
+   tilbert - Tilbert transform:         cs_diff(x,h,h)
+   itilbert - Inverse Tilbert transform: sc_diff(x,h,h)
+   hilbert - Hilbert transform:         cs_diff(x,inf,inf)
+   ihilbert - Inverse Hilbert transform: sc_diff(x,inf,inf)
+   cs_diff - cosh/sinh pseudo-derivative of periodic sequences
+   sc_diff - sinh/cosh pseudo-derivative of periodic sequences
+   ss_diff - sinh/sinh pseudo-derivative of periodic sequences
+   cc_diff - cosh/cosh pseudo-derivative of periodic sequences
+   shift - Shift periodic sequences
+
+Helper functions
+================
+
+.. autosummary::
+   :toctree: generated/
+
+   fftshift - Shift the zero-frequency component to the center of the spectrum
+   ifftshift - The inverse of `fftshift`
+   fftfreq - Return the Discrete Fourier Transform sample frequencies
+   rfftfreq - DFT sample frequencies (for usage with rfft, irfft)
+
+
+Convolutions (:mod:`scipy.fftpack.convolve`)
+============================================
+
+.. module:: scipy.fftpack.convolve
+
+.. autosummary::
+   :toctree: generated/
+
+   convolve
+   convolve_z
+   init_convolution_kernel
+   destroy_convolve_cache
+
+
+Other (:mod:`scipy.fftpack._fftpack`)
+=====================================
+
+.. module:: scipy.fftpack._fftpack
+
+.. autosummary::
+   :toctree: generated/
+
+   drfft
+   zfft
+   zrfft
+   zfftnd
+   destroy_drfft_cache
+   destroy_zfft_cache
+   destroy_zfftnd_cache
+
+"""
+
+__all__ = ['fft','ifft','fftn','ifftn','rfft','irfft',
+           'fft2','ifft2',
+           'diff',
+           'tilbert','itilbert','hilbert','ihilbert',
+           'sc_diff','cs_diff','cc_diff','ss_diff',
+           'shift',
+           'rfftfreq'
+           ]
 
 from fftpack_version import fftpack_version as __version__
