+++ scipy/integrate/__init__.py
@@ -1,8 +1,51 @@
-#
-# integrate - Integration routines
-#
+"""
+=============================================
+Integration and ODEs (:mod:`scipy.integrate`)
+=============================================
 
-from info import __doc__
+.. currentmodule:: scipy.integrate
+
+Integrating functions, given function object
+============================================
+
+.. autosummary::
+   :toctree: generated/
+
+   quad          -- General purpose integration.
+   dblquad       -- General purpose double integration.
+   tplquad       -- General purpose triple integration.
+   fixed_quad    -- Integrate func(x) using Gaussian quadrature of order n.
+   quadrature    -- Integrate with given tolerance using Gaussian quadrature.
+   romberg       -- Integrate func using Romberg integration.
+
+Integrating functions, given fixed samples
+==========================================
+
+.. autosummary::
+   :toctree: generated/
+
+   trapz         -- Use trapezoidal rule to compute integral from samples.
+   cumtrapz      -- Use trapezoidal rule to cumulatively compute integral.
+   simps         -- Use Simpson's rule to compute integral from samples.
+   romb          -- Use Romberg Integration to compute integral from
+                 -- (2**k + 1) evenly-spaced samples.
+
+.. seealso::
+
+   :mod:`scipy.special` for orthogonal polynomials (special) for Gaussian
+   quadrature roots and weights for other weighting factors and regions.
+
+Integrators of ODE systems
+==========================
+
+.. autosummary::
+   :toctree: generated/
+
+   odeint        -- General integration of ordinary differential equations.
+   ode           -- Integrate ODE using VODE and ZVODE routines.
+   complex_ode   -- Convert a complex-valued ODE to real-valued and integrate.
+
+"""
 
 from quadrature import *
 from odepack import *
