+++ scipy/io/__init__.py
@@ -1,9 +1,76 @@
-#
-# io - Data input and output
-#
+# -*- encoding:utf-8 -*-
+"""
+==================================
+Input and output (:mod:`scipy.io`)
+==================================
 
-from info import __doc__
+.. currentmodule:: scipy.io
 
+SciPy has many modules, classes, and functions available to read data
+from and write data to a variety of file formats.
+
+.. seealso:: :ref:`numpy-reference.routines.io` (in Numpy)
+
+MATLAB® files
+=============
+
+.. autosummary::
+   :toctree: generated/
+
+   loadmat - Read a MATLAB style mat file (version 4 through 7.1)
+   savemat - Write a MATLAB style mat file (version 4 through 7.1)
+
+Matrix Market files
+===================
+
+.. autosummary::
+   :toctree: generated/
+
+   mminfo - Query matrix info from Matrix Market formatted file
+   mmread - Read matrix from Matrix Market formatted file
+   mmwrite - Write matrix to Matrix Market formatted file
+
+Other
+=====
+
+.. autosummary::
+   :toctree: generated/
+
+   save_as_module - Data saved as module, accessed on load as attirbutes
+
+Wav sound files (:mod:`scipy.io.wavfile`)
+=========================================
+
+.. module:: scipy.io.wavfile
+
+.. autosummary::
+   :toctree: generated/
+
+   read
+   write
+
+Arff files (:mod:`scipy.io.arff`)
+=================================
+
+.. module:: scipy.io.arff
+
+.. autosummary::
+   :toctree: generated/
+
+   loadarff
+
+Netcdf (:mod:`scipy.io.netcdf`)
+===============================
+
+.. module:: scipy.io.netcdf
+
+.. autosummary::
+   :toctree: generated/
+
+   netcdf_file - A file object for NetCDF data
+   netcdf_variable - A data object for the netcdf module
+
+"""
 from numpy import deprecate
