--- scipy/lib/info.py
@@ -1,10 +1,0 @@
-"""
-Python wrappers to external libraries
-=====================================
-
-  lapack -- wrappers for LAPACK/ATLAS libraries
-  blas -- wrappers for BLAS/ATLAS libraries
-
-"""
-
-__all__ = ['lapack','blas']
