+++ scipy/spatial/__init__.py
@@ -1,8 +1,27 @@
-#
-# spatial - Distances
-#
+"""
+=============================================================
+Spatial algorithms and data structures (:mod:`scipy.spatial`)
+=============================================================
 
-from info import __doc__
+Nearest-neighbor queries:
+
+.. autosummary::
+   :toctree: generated/
+
+   KDTree      -- class for efficient nearest-neighbor queries
+   cKDTree     -- class for efficient nearest-neighbor queries (faster impl.)
+   distance    -- module containing many different distance measures
+
+Delaunay triangulation:
+
+.. autosummary::
+   :toctree: generated/
+
+   Delaunay
+   tsearch
+
+"""
+
 from kdtree import *
 from ckdtree import *
 from qhull import *
