--- scipy/spatial/info.py
@@ -1,25 +1,0 @@
-"""
-=============================================================
-Spatial algorithms and data structures (:mod:`scipy.spatial`)
-=============================================================
-
-Nearest-neighbor queries:
-
-.. autosummary::
-   :toctree: generated/
-
-   KDTree      -- class for efficient nearest-neighbor queries
-   cKDTree     -- class for efficient nearest-neighbor queries (faster impl.)
-   distance    -- module containing many different distance measures
-
-Delaunay triangulation:
-
-.. autosummary::
-   :toctree: generated/
-
-   Delaunay
-   tsearch
-
-"""
-
-postpone_import = 1
