--- scipy/weave/info.py
@@ -1,12 +1,0 @@
-"""
-C/C++ integration
-=================
-
-        inline     -- a function for including C/C++ code within Python
-        blitz      -- a function for compiling Numeric expressions to C++
-        ext_tools  -- a module that helps construct C/C++ extension modules.
-        accelerate -- a module that inline accelerates Python functions
-"""
-postpone_import = 1
-standalone = 1
-ignore = 1
